requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
dj-rest-auth==6.0.0
Django==5.1.3
django-allauth==65.0.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
idna==3.10
psycopg2-binary==2.9.6
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.9.0
